Rename ek::promise::operators to ek::promise::global_functions.
